feat#33: 게시글 조회, 수정 기능 구현 (#34)
* feat: 조회 조건 및 page size 정의
* feat: 게시글 목록 검색, 필터링 조회 기능 구현
* test: 게시글 목록 조회 테스트

* feat: 타인 프로필 조회 시 등록 게시글 조회 기능 구현
* test: 타인 등록 게시글 조회 관련 테스트 수정 및 통합 테스트 Mocking 제거

* feat: 게시글 상세 조회 기능 구현
* test: 게시글 상세 조회 기능 테스트
* feat: 게시글 상세 조회 시 로그인 된 사용자는 jwt 토큰 검증 및 인증 객체 저장

* feat: 내 게시글 목록 조회 기능 구현
* test: 내 게시글 목록 조회 기능 테스트

* feat: 게시글 수정 기능 구현
* test: 게시글 수정 기능 테스트

+++ src/test/unit/java/com/bob/domain/category/service/reader/CategoryReaderTest.java
@@ -47,7 +47,7 @@ class CategoryReaderTest {
   @DisplayName("카테고리 ID로 조회 - 존재하지 않는 경우 예외 발생")
   void 카테고리ID로_조회_시_존재하지_않으면_예외를_던진다() {
     // given
-    Long categoryId = -1L;
+    Integer categoryId = -1;
     given(categoryRepository.findById(categoryId)).willReturn(Optional.empty());
 
     // when & then
